Updated Collection For Demo
Since producing a youtube demo of this, I have modified the collection to
better organize what is delivered.  First, the network-topology examples
were removed since they were extraneous to a simple demo.  Next, some of
the requests were altered to adapt to changes in the upstream NETCONF
project, which hosts the RESTCONF implementation.

Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>